LoginGUI and Login and TheHotSpotsResort and ReadFromFile are good to go
Login and LoginGUI work and TheHotSpotsResort and ReadFromFile are all
working great.  An EIN and Password are pulled from textfeilds when a
button is clicked on the LoginGUI it then goes to the Login class and
test for length then the EIN and Password are checked by reading a text
file and comparing each line of stored passwords. We are using .txt
files as our databases. There are seperate files for our Admin and Staff
and a employees ID number specifies weather or not they are the Admin.
I have started a new Jframe class named BookCancelLookUpGUI which will
be the first screen. It will has 3 buttons. Book and Cancel and Look Up.
Book and Cancel will need a user to login, but look up will not