add a bunch of classes and components.
I added a bunch of classes, but currently don't have them in order. AS
we haven't discussed proper implemenation odrder, but  I have ideas I
can show everyone when we have our meeting this week. But the main
additions are the DatePicker class which has a  text box and a dropdown
arrow you and it lets you select one date. I think this option will be
good for printing reports. The big class i add implements a Jframe which
has 4 months(we can add as many rows and columns as we want. But I have
set it up to where we can select a span of dates. Which are saved in 2
different instance variables 1. a DateSpan 2.  a Calendar StartDate and
Calendar EndDate which are static and can be called from anywhere in the
package. I am not sure I f I am  using this properly. We will see