Updates to the main in TheHotSpotsResort. class
At this point we need to decide on what to do after the users logs in.
Does it go to a new frame or does where the user can click a button for
each room. Then after that it should open a calendar for the selected
room or should it open a calendar for each room this way the employee
can see what rooms are available during the next few months