Changed BookCancelLookUpGUI to 2 seperate classes StaffGUI and AdminGUI. I also changed the order the GUI's are called
I refactored and encapsulated many of these classes attributes and
changed them from protected to private and created getters and setters
for them. I also copied the old BOOKCancleLookUpGUI and made an AdminGUI
and StaffGUI from it. The purpose of this is that the AdminGUI will have
a new button printReportButton along with BookButton and Cancel Button
that the StaffGUI will have. This way when an Admin logs in they have
all the components they will need.   So far the order of classes called
is.  LoginGUI which has JTextField and JPasswordField and a LoginButton
if the loginButton is clicked it calls the Login class which verifies
user input in the 2 fields. if they are correct it opens either StaffGUI
or AdminGUI depending on who logs in. Then they will need to call the
prespective  GUIS based on user input.