Added a roomGUI with action listeners with all the buttons for rooms. Each button opens a new MyJXMonthViewCalendar
Add room GUI with action listeners with all the buttons for rooms. Each
button opens a new MyJXMonthViewCalendar. ON the month view calendars
you the user can select a span of dates. In this version the span is
only printed to the console and I havent implemented any form of a data
base. That should be coming in the next commit or next 2 commits. The
database will be .txt file