Updated The staff book a reservation path and added the look up.
Updated The staff book a reservation path and added the look up function
to the loginGUI class so when the lookUpJButton is clicked on the Login
JFrame it opens up a new JFrame and allow user to look in all room text
files for matching first name, last name, and corfirmation number. They
can search with any combination of the 3 or just one of them. The search
works fine but it prints all data from text file so I will need to debug
it later to find out why it copies all data from the file and puts the
data into a a jtable which is like an excel spreadsheet.  Alot of other
edits. I am also starting to work on getting the calendar for each room.
So the user will be able to see all the date that it is booked.