Made JXMonthView populate with all reservations and made those dates unselectable
Made JXMonthView populate with all reservations and made those dates
unselectable.  Set flagged dates which shows that the dates are already
reserved. Also made those dates unselectable. Except for the fact that
if the user selects a date before and selects all the way to a date
after the date span doesn't recognize that those dates are unselectable.