Print Report Added and debuged
Admin Print Report had been added. You select a date range and rooms you
want inclued in report and the action listener automatically addes it to
JTextArea on the form. Each action update the form. Then if user/Admin
wants to print there is a print report button. When this is selected a
joptionpane is opened telling user/Admin that report has been sent to
printer.