GUEST INFO GUI BUG FIXES number 3 update of jar for mac and pc to center properly
Bugs fixed on saveINfor when user hits cancel on joption pane. The form
would update if user changed text in text fileds it is now fixed it also
was allowing for special char after you hit cancel but they would get
saved now they can't even type them. It was a glitch as the key listener
was set up to consume these keys, but somehow after cancel was hit the
from was revalidating.. This problems has been resolved and no other
faults are known at this time, but we are still doing some pre -
acceptance testing.