System updated and new build Bugs fixed on JXMonthView Calendar
Bugs fixed last bug on the JXmonthView calendar has been corrected the
month was off by one in the JXMonthViewCalendar class method numOfDays.
It counts the number of days in a new reservation. This bug caused the
cost to be off for a reservation because it was counting days in next
month. It was a fault not a glitch as first thought